#8 Moving AccumulateTestRunner to test packages.
